Refactor to move handlers into server directory.
